implement two run loops
