+++ .run/List All nScopes.run.xml
@@ -1,12 +1,19 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="List All nScopes" type="CargoCommandRunConfiguration" factoryName="Cargo Command" folderName="Examples">
-    <option name="channel" value="DEFAULT" />
     <option name="command" value="run --example list_all_nscopes" />
-    <option name="allFeatures" value="false" />
-    <option name="emulateTerminal" value="false" />
-    <option name="backtrace" value="SHORT" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
-    <envs />
+    <option name="emulateTerminal" value="false" />
+    <option name="channel" value="DEFAULT" />
+    <option name="requiredFeatures" value="true" />
+    <option name="allFeatures" value="false" />
+    <option name="withSudo" value="false" />
+    <option name="buildTarget" value="REMOTE" />
+    <option name="backtrace" value="SHORT" />
+    <envs>
+      <env name="RUST_LOG" value="trace" />
+    </envs>
+    <option name="isRedirectInput" value="false" />
+    <option name="redirectInputPath" value="" />
     <method v="2">
       <option name="CARGO.BUILD_TASK_PROVIDER" enabled="true" />
     </method>
